requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.4
django-filter==22.1
django-multiupload==0.6.1
docopt==0.6.2
idna==3.3
Pillow==9.1.1
pipreqs==0.4.11
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
yarg==0.1.9
